Update the structure of the project
